+++ src/main/java/com/example/todolist_prac/controller/TodoController.java
@@ -1,10 +1,11 @@
 package com.example.todolist_prac.controller;
 
-import com.example.todolist_prac.model.TodoEntity;
 import com.example.todolist_prac.model.TodoRequest;
 import com.example.todolist_prac.model.TodoResponse;
 import com.example.todolist_prac.service.TodoService;
+import com.example.todolist_prac.service.TodoServiceImpl;
 import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,17 +14,15 @@ import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Slf4j
 //@CrossOrigin
-@AllArgsConstructor
+@RequiredArgsConstructor
 @RestController
 @RequestMapping("/todo")
 public class TodoController {
 
-    @Autowired
-    private final TodoService service;
+    private final TodoService todoService;
 
     @PostMapping
     public ResponseEntity<TodoResponse> create(@RequestBody TodoRequest request) {
@@ -41,7 +40,7 @@ public class TodoController {
             request.setCompleted(false);
         }
 
-        TodoResponse todoResponse = this.service.add(request);
+        TodoResponse todoResponse = todoService.add(request);
 
         return new ResponseEntity<>(todoResponse, HttpStatus.CREATED);
     }
@@ -49,7 +48,7 @@ public class TodoController {
     @GetMapping("{id}")
     public ResponseEntity<TodoResponse> readOne(@PathVariable Long id) {
         log.info("Read One");
-        TodoResponse todoResponse = this.service.searchById(id);
+        TodoResponse todoResponse = todoService.searchById(id);
 
 //        return new ResponseEntity<>(todoResponse, HttpStatus.OK);
         return ResponseEntity.ok(todoResponse);
@@ -58,34 +57,34 @@ public class TodoController {
     @GetMapping
     public List<TodoResponse> readAll() {
         log.info("Read All");
-        return this.service.searchAll();
+        return todoService.searchAll();
     }
 
     @PutMapping("{id}")
     public ResponseEntity<TodoResponse> update(@PathVariable Long id) {
         log.info("Update");
-        TodoResponse todoResponse = this.service.updateById(id);
+        TodoResponse todoResponse = todoService.updateById(id);
         return new ResponseEntity<>(todoResponse, HttpStatus.OK);
     }
 
 /*    @PutMapping("{id}")
     public ResponseEntity<TodoResponse> update2(@PathVariable Long id, @RequestBody TodoRequest request) {
         log.info("Update2");
-        TodoResponse todoResponse = this.service.updateById(id, request);
+        TodoResponse todoResponse = todoService.updateById(id, request);
         return new ResponseEntity<>(todoResponse, HttpStatus.OK);
     }*/
 
     @DeleteMapping("{id}")
     public ResponseEntity<String> deleteOne(@PathVariable Long id) {
         log.info("Delete");
-        this.service.deleteById(id);
+        todoService.deleteById(id);
         return new ResponseEntity<>("deleteOne successfully.", HttpStatus.OK);
     }
 
     @DeleteMapping
     public ResponseEntity<?> deleteAll() {
         log.info("Delete All");
-        this.service.deleteAll();
+        todoService.deleteAll();
         return new ResponseEntity<>("deleteAll successfully.", HttpStatus.OK);
     }
