+++ src/main/java/com/example/todolist_prac/model/TodoEntity.java
@@ -1,9 +1,6 @@
 package com.example.todolist_prac.model;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
